test addition and taking some code from PlanListActivityTest out to ChildViewClicker, as I need it in TaskListActivityTest and it would be redundant
