+++ Friendly-plans/app/src/androidTest/java/pg/autyzm/friendly_plans/manager_app/PlanListActivityTest.java
@@ -21,6 +21,7 @@ import java.util.List;
 import database.entities.PlanTemplate;
 import database.repository.PlanTemplateRepository;
 import pg.autyzm.friendly_plans.R;
+import pg.autyzm.friendly_plans.resource.ChildViewClicker;
 import pg.autyzm.friendly_plans.resource.DaoSessionResource;
 import pg.autyzm.friendly_plans.manager_app.view.plan_list.PlanListActivity;
 
@@ -124,7 +125,7 @@ public class PlanListActivityTest {
         onView(withId(R.id.rv_plan_list))
                 .perform(RecyclerViewActions
                         .actionOnItemAtPosition(testedTaskPosition,
-                                clickChildViewWithId(R.id.id_remove_plan)));
+                                new ChildViewClicker(R.id.id_remove_plan)));
         onView(withId(R.id.rv_plan_list)).perform(scrollToPosition(testedTaskPosition));
         onView(withRecyclerView(R.id.rv_plan_list)
                 .atPosition(testedTaskPosition))
@@ -142,7 +143,7 @@ public class PlanListActivityTest {
         onView(withId(R.id.rv_plan_list))
                 .perform(RecyclerViewActions
                         .actionOnItemAtPosition(0,
-                                clickChildViewWithId(R.id.id_remove_plan)));
+                                new ChildViewClicker(R.id.id_remove_plan)));
         closeSoftKeyboard();
 
         onView(withRecyclerView(R.id.rv_plan_list)
@@ -159,7 +160,7 @@ public class PlanListActivityTest {
         onView(withId(R.id.rv_plan_list))
                 .perform(RecyclerViewActions
                         .actionOnItemAtPosition(0,
-                                clickChildViewWithId(R.id.id_remove_plan)));
+                                new ChildViewClicker(R.id.id_remove_plan)));
         onView(isAssignableFrom(EditText.class)).perform(clearText());
         closeSoftKeyboard();
 
@@ -193,25 +194,5 @@ public class PlanListActivityTest {
         List<PlanTemplate> planTemplates = planTemplateRepository.get(newName);
         assertThat(planTemplates.size(), is(1));
 
-    }
-
-    public static ViewAction clickChildViewWithId(final int id) {
-        return new ViewAction() {
-            @Override
-            public Matcher<View> getConstraints() {
-                return null;
-            }
-
-            @Override
-            public String getDescription() {
-                return null;
-            }
-
-            @Override
-            public void perform(UiController uiController, View view) {
-                View v = view.findViewById(id);
-                v.performClick();
-            }
-        };
     }
 }
